requirements.txt
appdirs==1.4.4
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
distlib==0.3.1
filelock==3.0.12
hyperlink==21.0.0
identify==1.5.14
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
nodeenv==1.5.0
parsel==1.6.0
Pillow==8.1.2
pre-commit==2.10.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dotenv==0.15.0
PyYAML==5.4.1
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.3
virtualenv==20.4.2
w3lib==1.22.0
zope.interface==5.2.0
